% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{colors}
\alias{colors}
\title{Image plotter}
\usage{
colors(X, Y)
}
\arguments{
\item{X, Y}{}
}
\value{
returns a dataframe of images
}
\description{
Function to extract image colors
}
\examples{
colors() # This is unclear
}
