% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{edge_analysis}
\alias{edge_analysis}
\title{Edge analysis}
\usage{
edge_analysis(images)
}
\arguments{
\item{x}{Folder where images are stored}
}
\value{
Returns a dataframe consisting of images, PQ, ST
}
\description{
Performs edge analysis
}
\details{
this function uses the same 16 cell grid for image segmentation used in the symmetry function

underlying math in this function is from a mean of canny edges detected per cell and deviation of them

the mean would speak to the total value of "ink" in the zone and the deviation may inform the character of the edges
}
\examples{
load_images(here("Images/"))
}
