% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fluency}
\alias{fluency}
\title{Fluency}
\usage{
fluency(images)
}
\arguments{
\item{images}{Image to be read in}
}
\value{
Returns image contrast, similarity, symmetry, complexity
}
\description{
Get image fluency
}
\examples{
fluency(here("Images/image_1.png"))
}
