% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fluency_analysis}
\alias{fluency_analysis}
\title{Fluency}
\usage{
fluency_analysis(image)
}
\arguments{
\item{image}{Image to be read in}
}
\value{
Returns image contrast, similarity, symmetry, complexity
}
\description{
Get image fluency
}
\details{
this function implements multiple tests from the Imagefluency package returning a single dataframe, this function can take quite a while to run.
}
\examples{
fluency(here("Images/image_1.png"))
}
