% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{imageplot_output}
\alias{imageplot_output}
\title{function that allows you to pass alpha to a GG plot that also encodes other things}
\usage{
imageplot_output(D, X, Y, A, S)
}
\arguments{
\item{D}{is where the data is}

\item{X}{is the X var}

\item{Y}{is the Y var}

\item{A}{is the alpha}

\item{S}{is relative size}
}
\description{
function that allows you to pass alpha to a GG plot that also encodes other things
}
