% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lower_converter}
\alias{lower_converter}
\title{Lower converter}
\usage{
lower_converter(x)
}
\arguments{
\item{X}{Folder where images are stored}
}
\value{
returns a dataframe of images
}
\description{
Load all images into one
}
