% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{measure_images}
\alias{measure_images}
\title{Measure Image}
\usage{
measure_images(x)
}
\arguments{
\item{x}{Image to be read in}
}
\value{
Returns image information and text to console
}
\description{
Measure image information and ocr
}
\examples{
measure_images(here("Images/image_1.png"))
}
