% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot2}
\alias{plot2}
\title{Plot images}
\usage{
plot2(
  img,
  n = 10000,
  lower = c(0, 0.55, 0),
  upper = c(0.25, 1, 0.25),
  color.space = "rgb",
  ref.white = NULL,
  pch = 20,
  main = "default",
  from = "sRGB",
  xlim = "default",
  ylim = "default",
  zlim = "default",
  ...
)
}
\arguments{
\item{img, n=1000, lower=c(0, 0.55, 0), upper=c(0.25, 1, 0.25), color.space="rgb", ref.white=NULL, pch=20, main="default", from="sRGB", xlim="default", ylim="default", zlim="default", ...}{}
}
\value{
returns a plot
}
\description{
Plotting function
}
\examples{
plot2(here("Images/"))
}
