% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{edge_analysis}
\alias{edge_analysis}
\title{Edge analysis}
\usage{
edge_analysis(x)
}
\arguments{
\item{x}{Folder where images are stored}
}
\value{
Returns a dataframe consisting of images, PQ, ST
}
\description{
Performs edge analysis
}
\examples{
load_images(here("Images/"))
}
